backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
Pillow>=9.5.0,<10.0.0
typing-extensions>=4.5.0
